+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **[BC]** Rename `AnyComparableInterface` to `AnyComparable`
 - **[BC]** Rename `ExtendedComparableInterface` to `ExtendedComparable`
 - **[BC]** Rename `SelfComparableInterface` to `SelfComparable`
+- **[BC]** Rename `SubClassComparableInterface` to `SubClassComparable`
 - **[BC]** Remove `AbstractExtendedComparable` class, use `ExtendedComparableTrait` instead
 
 ### 2.0.0 (2018-11-06)